Refactor #576: queue / system-webhook 系を handler_stubs から分離
admin package の handler_stubs.go から queue (12 endpoints + 10 helpers)
と system-webhook (6 endpoints) を thematic ファイルに切り出す。挙動は
完全に等価で、本コミットは pure file move + helper organization。

#574 (PR #581) / #575 (PR #582) で確立した handler_stubs 解体パターンを
踏襲。中身は既に本実装済みで、ファイル名のみが misleading な "stub"
だったため、disassembly により discoverability を改善する。

queue.go:
- 12 handlers: Clear / DeliverDelayed / InboxDelayed / Jobs / PromoteJobs /
  QueueStats / Queues / RemoveJob / RetryJob / ShowJob / ShowJobLogs / Stats
- 10 helpers: listDelayedTasks / parseStateField / listTasksForState /
  shapeQueueForFrontend / metricsToFrontend / fetchQueueMetrics /
  packTaskSummary / formatUnixMillisOrZero / stacktraceFrom / rawJSONOrString
- delayedTasksMaxFetch const

system_webhook.go:
- 6 handlers: Create / Delete / List / Show / Test / Update

handler_stubs.go side-effects:
- 704 行削除 (2139 -> 1435)
- 未使用化した import を削除: encoding/json, strings

Modernize: listDelayedTasks の end clamp を min() に変更
(go vet modernize lint への対応)。

Coverage: 84.1% -> 84.1% (admin pkg 閾値 80% を上回る、regression なし)。
テストファイル (queue_stats_test.go, system_webhook_test.go) は touch せず、
21 + 20 件のテストすべてグリーン。

将来 cleanup タスク:
- captcha tests が queue_stats_test.go に混入 + sendWebhookTest が email.go
  に同居 -> #583 で別 issue 化
- system_webhook_test.go の AbuseReportNotificationRecipient 混入 -> #577
  にコメント追記済

Closes #576
